Added submenu for simpler loading of pre-trained models
Added a small sub-routine that scans for previously
saved models. These can be loaded by simply selecting
the menu option instead of providing the full path.
The menu is presented in reverse chronological order
and the final option "Specify paths manually" preserves
the ability to provide a full path for the checkpoint
and replay buffer as was done previously.