repeat last reward after teminate state in make_target()
In the papar,
> 3) Terminal nodes. AlphaZero stopped the search at tree nodes representing terminal states and used the terminal value provided by the simulator instead of the value produced by the network. MuZero does not give
> special treatment to terminal nodes and always uses the value predicted by the network. Inside the tree, the
> search can proceed past a terminal node - in this case the network is expected to always predict the same
> value. This is achieved by treating terminal states as absorbing states during training

What is "the same value"?
In the current implemantation, after terminating game, values and rewards are 0.
In this new implementation, valus are 0, but rewards are last_reward value.

This way, when the node is expanded past the end state in MCTS, the reword is repeated, which makes the choice clearer and leads to good policy learning.