Update game logic and error handling in spiel.py
The commit introduces a new function, `get_observation_tensor_shape(game: Game) -> Tuple`, which handles game observation dimensions and improves error handling in this respect. The game in play was switched from tic_tac_toe to backgammon. Moreover, the error handling for the method `self.board.rewards()` was improved to prevent a loss of the entire operation when an exception is catched.